Refactor: Extract home sections into separate files
